Moved ping to seperate file, added csv support
