+++ examples/normalizer/test-units.csv
@@ -1,0 +1,3 @@
+r1,cisco
+r2,cisco
+r3,cisco
